app/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.6.20
cycler==0.10.0
distlib==0.3.1
Django==3.1.7
django-createsuperuserwithpassword==2.0.0
django-extensions==3.1.1
django-filter==21.1
django-mathfilters==1.0.0
filelock==3.0.12
kiwisolver==1.3.2
Markdown==3.2.2
matplotlib==3.4.3
numpy==1.20.2
pandas==1.2.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
thinkx==1.1.3
virtualenv==20.4.3
